+++ ledger/ledger-api-integration-tests/src/test/lib/scala/com/digitalasset/platform/apitesting/TestTemplateIdentifiers.scala
@@ -101,6 +101,10 @@ final case class TestTemplateIdentifiers(testPackageId: String) {
     Identifier(testPackageId, "Test.Divulgence2", "Test", "Divulgence2")
   val decimalRounding =
     Identifier(testPackageId, "Test.DecimalRounding", "Test", "DecimalRounding")
+  val delegated = Identifier(testPackageId, moduleName = "Test", entityName = "Delegated")
+  val delegation = Identifier(testPackageId, moduleName = "Test", entityName = "Delegation")
+  val showDelegated =
+    Identifier(testPackageId, moduleName = "Test", entityName = "ShowDelegated")
   val allTemplates =
     List(
       dummy,