+++ ledger/ledger-api-integration-tests/src/test/semanticsuite/scala/com/digitalasset/platform/semantictest/SandboxSemanticTestsLfRunner.scala
@@ -30,6 +30,11 @@ class SandboxSemanticTestsLfRunner
     .withDarFile(darFile.toPath)
     .withTimeProvider(TimeProviderType.StaticAllowBackwards)
 
+  // TODO SC delete when implicit disclosure supplied in PostgresLedgerDao
+  override protected def fixtureIdsEnabled
+    : Set[com.digitalasset.platform.apitesting.LedgerBackend] =
+    Set(com.digitalasset.platform.apitesting.LedgerBackend.SandboxInMemory)
+
   lazy val (mainPkgId, packages) = {
     val dar = UniversalArchiveReader().readFile(darFile).get
     val packages = Map(dar.all.map {